Remove need to mark methods with safer_ffi_gen
This allows to support cfg_attr. cfg_attr on methods would not work because safer_ffi_gen in that position is not an attribute; it is an arbitrary token consumed/understood exclusively by our macro.

The drawback of this change is that impl blocks that don't expose all their methods through FFI need to be split.