Fix type generic over lifetime (#74)
* Don't use unnamed lifetime in return type

If the type code is generated for is generic over lifetimes, multiple lifetimes are involved and named lifetimes are required.

* Fix generated code for `Type<'a>`

Somehow Rust does not always infer `'a: 'b` for `&'b Foo<&'a Bar>`. The fix consists in always adding this constraint explicitly.

* Fix warning about resolver

* Temporarily ignore clippy warning

Until `safer-ffi` fixes the code it generates.

* Fix lifetime error with headers feature

The code generated by safer-ffi with the headers feature did not compile without specifying this constraint on the original function. This is another instance of `'a: 'b` not being inferred for `&'b Foo<&'a Bar>`.

* Solve lifetime woes in generated clone function for type generic over lifetime